requirements.txt
sqlalchemy
psycopg2
python-dotenv
requests